Revert "Revert "Fix:/#706 all bugs in episode details screen""

+++ ui/src/main/java/com/baghdad/ui/feature/tvShowDetails/component/EpisodesCard.kt
@@ -34,7 +34,7 @@ fun EpisodeCard(
     episodeName: String,
     releaseDate: String,
     duration: String,
-    rating: Double,
+    rating: Int,
     modifier: Modifier = Modifier,
 ) {
     Row(
